interest_agent/requirements.txt
anthropic>=0.40.0
feedparser>=6.0.0
ddgs>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
# Profile Builder 依賴
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
